Added missing @DrawableRes to Resources#openRawResource() and Resources#openRawResourceFd().
According to javadoc: "This can only be
used with resources whose value is the name of an asset file -- that is,
it can be used to open drawable, sound, and raw resources". So drawable
should also be allowed (and it really works).

Corrected corresponding methods and exception messages in ResourcesImpl
which mentioned only drawable but not raw.

Change-Id: Iba81e9ef5ce412d33843b820a40d949fd7aea7fc
(cherry picked from commit 685b5ad8abdf227a830763bd543754a8b7cfc00f)